reuse test logic
